@@ -1,7 +1,7 @@
 {
   "name": "code-oss-dev",
   "version": "1.93.0",
-  "distro": "47c05c9902553fc1733d36ded737146d7685ad5a",
+  "distro": "02ac83d21dd98da723344443f9a26fdbb4cd9098",
   "author": {
     "name": "Microsoft Corporation"
   },
@@ -155,7 +155,7 @@
     "cssnano": "^6.0.3",
     "debounce": "^1.0.0",
     "deemon": "^1.8.0",
-    "electron": "30.3.1",
+    "electron": "30.4.0",
     "eslint": "8.36.0",
     "eslint-plugin-header": "3.1.1",
     "eslint-plugin-jsdoc": "^46.5.0",
