@@ -3,11 +3,12 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { Button } from 'vs/base/browser/ui/button/button';
-import { IAction } from 'vs/base/common/actions';
+import { Button, ButtonWithDropdown } from 'vs/base/browser/ui/button/button';
+import { ActionRunner, IAction } from 'vs/base/common/actions';
 import { DisposableStore, IDisposable } from 'vs/base/common/lifecycle';
 import { IMenu } from 'vs/platform/actions/common/actions';
 import { IContextKeyService } from 'vs/platform/contextkey/common/contextkey';
+import { IContextMenuService } from 'vs/platform/contextview/browser/contextView';
 import { IKeybindingService } from 'vs/platform/keybinding/common/keybinding';
 import { defaultButtonStyles } from 'vs/platform/theme/browser/defaultStyles';
 import { CommentCommandId } from 'vs/workbench/contrib/comments/common/commentCommandIds';
@@ -20,6 +21,7 @@ export class CommentFormActions implements IDisposable {
 	constructor(
 		private readonly keybindingService: IKeybindingService,
 		private readonly contextKeyService: IContextKeyService,
+		private readonly contextMenuService: IContextMenuService,
 		private container: HTMLElement,
 		private actionHandler: (action: IAction) => void,
 		private readonly maxActions?: number
@@ -34,16 +36,30 @@ export class CommentFormActions implements IDisposable {
 		const groups = menu.getActions({ shouldForwardArgs: true });
 		let isPrimary: boolean = !hasOnlySecondaryActions;
 		for (const group of groups) {
-			const [, actions] = group;
+			const [groupId, actions] = group;
 
 			this._actions = actions;
-			for (const action of actions) {
+			for (const current of groupId === 'inline' ? actions : [actions]) {
+				const [action, dropDownActions] = Array.isArray(current) ? [current[0], current.slice(1)] : [current, []];
 				let keybinding = this.keybindingService.lookupKeybinding(action.id, this.contextKeyService)?.getLabel();
 				if (!keybinding && isPrimary) {
 					keybinding = this.keybindingService.lookupKeybinding(CommentCommandId.Submit, this.contextKeyService)?.getLabel();
 				}
 				const title = keybinding ? `${action.label} (${keybinding})` : action.label;
-				const button = new Button(this.container, { secondary: !isPrimary, title, ...defaultButtonStyles });
+				const actionHandler = this.actionHandler;
+				const button = dropDownActions.length ? new ButtonWithDropdown(this.container, {
+					contextMenuProvider: this.contextMenuService,
+					actions: dropDownActions,
+					actionRunner: new class extends ActionRunner {
+						protected override async runAction(action: IAction, context?: unknown): Promise<void> {
+							return actionHandler(action);
+						}
+					},
+					secondary: !isPrimary,
+					title,
+					addPrimaryActionToDropdown: false,
+					...defaultButtonStyles
+				}) : new Button(this.container, { secondary: !isPrimary, title, ...defaultButtonStyles });
 
 				isPrimary = false;
 				this._buttonElements.push(button.element);
