@@ -42,13 +42,14 @@ import { InlineChatContentWidget } from 'vs/workbench/contrib/inlineChat/browser
 import { HunkInformation, Session, StashedSession } from 'vs/workbench/contrib/inlineChat/browser/inlineChatSession';
 import { InlineChatError } from 'vs/workbench/contrib/inlineChat/browser/inlineChatSessionServiceImpl';
 import { EditModeStrategy, HunkAction, IEditObserver, LiveStrategy, PreviewStrategy, ProgressingEditsOptions } from 'vs/workbench/contrib/inlineChat/browser/inlineChatStrategies';
-import { CTX_INLINE_CHAT_EDITING, CTX_INLINE_CHAT_REQUEST_IN_PROGRESS, CTX_INLINE_CHAT_RESPONSE_TYPE, CTX_INLINE_CHAT_SUPPORT_REPORT_ISSUE, CTX_INLINE_CHAT_USER_DID_EDIT, CTX_INLINE_CHAT_VISIBLE, EditMode, INLINE_CHAT_ID, InlineChatConfigKeys, InlineChatResponseType } from 'vs/workbench/contrib/inlineChat/common/inlineChat';
+import { CTX_INLINE_CHAT_EDITING, CTX_INLINE_CHAT_REQUEST_IN_PROGRESS, CTX_INLINE_CHAT_RESPONSE_TYPE, CTX_INLINE_CHAT_USER_DID_EDIT, CTX_INLINE_CHAT_VISIBLE, EditMode, INLINE_CHAT_ID, InlineChatConfigKeys, InlineChatResponseType } from 'vs/workbench/contrib/inlineChat/common/inlineChat';
 import { INotebookEditorService } from 'vs/workbench/contrib/notebook/browser/services/notebookEditorService';
 import { IEditorService, SIDE_GROUP } from 'vs/workbench/services/editor/common/editorService';
 import { IViewsService } from 'vs/workbench/services/views/common/viewsService';
 import { IInlineChatSavingService } from './inlineChatSavingService';
 import { IInlineChatSessionService } from './inlineChatSessionService';
 import { InlineChatZoneWidget } from './inlineChatZoneWidget';
+import { CONTEXT_RESPONSE, CONTEXT_RESPONSE_ERROR } from 'vs/workbench/contrib/chat/common/chatContextKeys';
 
 export const enum State {
 	CREATE_SESSION = 'CREATE_SESSION',
@@ -113,7 +114,8 @@ export class InlineChatController implements IEditorContribution {
 	private readonly _ctxResponseType: IContextKey<undefined | InlineChatResponseType>;
 	private readonly _ctxUserDidEdit: IContextKey<boolean>;
 	private readonly _ctxRequestInProgress: IContextKey<boolean>;
-	private readonly _ctxSupportsReportIssue: IContextKey<boolean>;
+
+	private readonly _ctxResponse: IContextKey<boolean>;
 
 	private readonly _messages = this._store.add(new Emitter<Message>());
 	protected readonly _onDidEnterState = this._store.add(new Emitter<State>());
@@ -154,7 +156,9 @@ export class InlineChatController implements IEditorContribution {
 		this._ctxUserDidEdit = CTX_INLINE_CHAT_USER_DID_EDIT.bindTo(contextKeyService);
 		this._ctxResponseType = CTX_INLINE_CHAT_RESPONSE_TYPE.bindTo(contextKeyService);
 		this._ctxRequestInProgress = CTX_INLINE_CHAT_REQUEST_IN_PROGRESS.bindTo(contextKeyService);
-		this._ctxSupportsReportIssue = CTX_INLINE_CHAT_SUPPORT_REPORT_ISSUE.bindTo(contextKeyService);
+
+		this._ctxResponse = CONTEXT_RESPONSE.bindTo(contextKeyService);
+		CONTEXT_RESPONSE_ERROR.bindTo(contextKeyService);
 
 		this._ui = new Lazy(() => {
 
@@ -487,6 +491,8 @@ export class InlineChatController implements IEditorContribution {
 			const diff = await this._editorWorkerService.computeDiff(this._session.textModel0.uri, this._session.textModelN.uri, { computeMoves: false, maxComputationTimeMs: Number.MAX_SAFE_INTEGER, ignoreTrimWhitespace: false }, 'advanced');
 			this._session.wholeRange.fixup(diff?.changes ?? []);
 			await this._session.hunkData.recompute(editState, diff);
+
+			this._updateCtxResponseType();
 		}
 		options.position = await this._strategy.renderChanges();
 
@@ -574,8 +580,6 @@ export class InlineChatController implements IEditorContribution {
 
 		assertType(request);
 		assertType(request.response);
-
-		this._ctxSupportsReportIssue.set(request.response.agent?.metadata.supportIssueReporting ?? false);
 
 		this._showWidget(this._session.headless, false);
 		this._ui.value.zone.widget.selectAll(false);
@@ -928,6 +932,7 @@ export class InlineChatController implements IEditorContribution {
 			}
 		}
 		this._ctxResponseType.set(responseType);
+		this._ctxResponse.set(responseType !== InlineChatResponseType.None);
 	}
 
 	private _createChatTextEditGroupState(): IChatTextEditGroupState {
